No me acuerdo que subo

+++ ClinicaFrba-BETTER_CALL_JUAN/ClinicaFrba/Modelo/Objetos/PlanMedico.cs
@@ -8,7 +8,7 @@ namespace ClinicaFrba
 {
     class PlanMedico
     {
-        public int codigo { get; set; }
+        public decimal codigo { get; set; }
         public string descripcion { get; set; }
         public Decimal precioBonoConsulta { get; set; }
         public Decimal precioBonoFarmacia { get; set; }
